@@ -16,7 +16,6 @@ package transferer
 
 import (
 	"bytes"
-	"context"
 	"testing"
 	"time"
 
@@ -24,7 +23,7 @@ import (
 )
 
 func TestExport(t *testing.T) {
-	ctx := context.Background()
+	ctx := t.Context()
 
 	type lsOpts struct {
 		status          LoadStatus
@@ -127,9 +126,12 @@ func TestExport(t *testing.T) {
 					sources:      sources,
 					workersCount: opt.workersCount,
 					file:         bytes.NewBuffer(nil),
-					encrypted:    newFalse(),
-					key:          newString(),
-					iv:           newString(),
+				}
+				e := Encyptor{
+					noEncryption: true,
+					justKey:      false,
+					pass:         "",
+					filepath:     "",
 				}
 				var meta dump.Meta
 				var chunks []dump.ChunkMeta
@@ -144,7 +146,7 @@ func TestExport(t *testing.T) {
 				if err != nil {
 					t.Fatal(err, "failed to create new chunk pool")
 				}
-				err = tr.Export(ctx, fakeStatusGetter{status: tt.loadStatus.status, waitCount: tt.loadStatus.waitCount, statusAfterWait: tt.loadStatus.statusAfterWait, count: new(int)}, meta, pool, new(bytes.Buffer), false, false)
+				err = tr.Export(ctx, fakeStatusGetter{status: tt.loadStatus.status, waitCount: tt.loadStatus.waitCount, statusAfterWait: tt.loadStatus.statusAfterWait, count: new(int)}, meta, pool, new(bytes.Buffer), e)
 				if err != nil {
 					if tt.shouldErr {
 						return
