@@ -32,12 +32,9 @@ type Transferer struct {
 	sources      []dump.Source
 	workersCount int
 	file         io.ReadWriter
-	encrypted    *bool
-	key          *string
-	iv           *string
 }
 
-func New(file io.ReadWriter, s []dump.Source, workersCount int, encrypted *bool, key, iv *string) (*Transferer, error) {
+func New(file io.ReadWriter, s []dump.Source, workersCount int) (*Transferer, error) {
 	if len(s) == 0 {
 		return nil, errors.New("no sources provided")
 	}
@@ -50,9 +47,6 @@ func New(file io.ReadWriter, s []dump.Source, workersCount int, encrypted *bool,
 		sources:      s,
 		workersCount: workersCount,
 		file:         file,
-		encrypted:    encrypted,
-		key:          key,
-		iv:           iv,
 	}, nil
 }
 
@@ -70,5 +64,5 @@ func (t Transferer) sourceByType(st dump.SourceType) (dump.Source, bool) { //nol
 			return s, true
 		}
 	}
-	return nil, false
+	return nil, false //nolint:ireturn
 }
